+++ test/project/html-sketchapp.config.js
@@ -1,13 +1,14 @@
 // https://github.com/seek-oss/html-sketchapp-cli#viewport-sizes-and-responsive-design
-const breakpoints = require('./src/lib/breakpoints.json')
-const viewports = Object.keys(breakpoints).reduce((result, name) => {
-  const width = breakpoints[name]
-  result[name] = `${width}x2048`
+const viewports = require('./src/lib/viewports.json')
+
+const viewportsSketch = Object.keys(viewports).reduce((result, name) => {
+  const { width, height } = viewports[name]
+  result[name] = `${width}x${height}`
   return result
 }, {})
 
 module.exports = {
   url: 'http://localhost:3000/_sketch.html',
   outDir: '../tmp/_sketch',
-  viewports
+  viewports: viewportsSketch
 }